requirements.txt
Flask
google-generativeai
python-dotenv
PyPDF2
python-docx
fpdf2
gunicorn
pytesseract
Pillow  